requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
# FastAPI 0.111 expects Pydantic v2
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.0
pytest==8.2.2
pytest-cov==5.0.0
SQLAlchemy==2.0.32
alembic==1.13.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
structlog==24.1.0
webdavclient3==3.14.6
python-dotenv==1.0.1
